+++ src/vs/platform/dialogs/common/dialogs.ts
@@ -8,8 +8,12 @@ import { createDecorator } from 'vs/platform/instantiation/common/instantiation'
 import { URI } from 'vs/base/common/uri';
 import { basename } from 'vs/base/common/resources';
 import { localize } from 'vs/nls';
-import { FileFilter } from 'vs/platform/windows/common/windows';
 import { ITelemetryData } from 'vs/platform/telemetry/common/telemetry';
+
+export interface FileFilter {
+	extensions: string[];
+	name: string;
+}
 
 export type DialogType = 'none' | 'info' | 'error' | 'question' | 'warning';
